+++ CHANGES.rst
@@ -1,5 +1,14 @@
 Changelog
 =========
+2.1.dev0
+--------
+- Compatibility with Chameleon
+
+  Chameleon tries to translate everything that is not a string or number or
+  possesses a ``__html__`` attribute. To regain compatibility, we reject any
+  message that is not a string.
+
+  See `Zope#876 <https://github.com/zopefoundation/Zope/issues/876>`_
 
 2.0 (2020-07-02)
 ----------------